+++ include/Qt/PlayerPrivate.h
@@ -47,15 +47,15 @@ namespace openshot
     class PlayerPrivate : Thread
     {
     tr1::shared_ptr<Frame> frame; /// The current frame
-	int video_position; /// The current frame position.
-	int audio_position; /// The current frame position.
+	long int video_position; /// The current frame position.
+	long int audio_position; /// The current frame position.
 	ReaderBase *reader; /// The reader which powers this player
 	AudioPlaybackThread *audioPlayback; /// The audio thread
 	VideoPlaybackThread *videoPlayback; /// The video thread
 	VideoCacheThread *videoCache; /// The cache thread
 	int speed; /// The speed and direction to playback a reader (1=normal, 2=fast, 3=faster, -1=rewind, etc...)
 	RendererBase *renderer;
-	int last_video_position; /// The last frame actually displayed
+	long int last_video_position; /// The last frame actually displayed
 
 	/// Constructor
 	PlayerPrivate(RendererBase *rb);
