+++ include/Qt/VideoCacheThread.h
@@ -45,8 +45,8 @@ namespace openshot
 	std::tr1::shared_ptr<Frame> frame;
 	int speed;
 	bool is_playing;
-	int position;
-	int current_display_frame;
+	long int position;
+	long int current_display_frame;
 	ReaderBase *reader;
 	int max_frames;
 
@@ -56,7 +56,7 @@ namespace openshot
 	~VideoCacheThread();
 
 	/// Get the currently playing frame number (if any)
-	int getCurrentFramePosition();
+	long int getCurrentFramePosition();
 
     /// Get Speed (The speed and direction to playback a reader (1=normal, 2=fast, 3=faster, -1=rewind, etc...)
     int getSpeed() const { return speed; }
@@ -65,10 +65,10 @@ namespace openshot
 	void Play();
 
 	/// Seek the reader to a particular frame number
-	void Seek(int new_position);
+	void Seek(long int new_position);
 
 	/// Set the currently displaying frame number
-	void setCurrentFramePosition(int current_frame_number);
+	void setCurrentFramePosition(long int current_frame_number);
 
     /// Set Speed (The speed and direction to playback a reader (1=normal, 2=fast, 3=faster, -1=rewind, etc...)
     void setSpeed(int new_speed) { speed = new_speed; }
